feat: support compile windows platform from source.

+++ bridge/bindings/qjs/js_event_handler.cc
@@ -76,7 +76,7 @@ void JSEventHandler::InvokeInternal(EventTarget& event_target, Event& event, Exc
     arguments.emplace_back(event.ToValue());
   }
 
-  ScriptValue result = event_handler_->Invoke(event.ctx(), event_target.ToValue(), arguments.size(), arguments.data());
+  ScriptValue result = event_handler_->Invoke(event.ctx(), event_target.ToValue(), (int) arguments.size(), arguments.data());
   if (result.IsException()) {
     exception_state.ThrowException(event.ctx(), result.QJSValue());
     return;
