+++ bridge/bindings/qjs/script_wrappable.h
@@ -57,7 +57,7 @@ class ScriptWrappable : public GarbageCollected<ScriptWrappable> {
   multi_threading::Dispatcher* GetDispatcher() const;
   FORCE_INLINE JSContext* ctx() const { return ctx_; }
   FORCE_INLINE JSRuntime* runtime() const { return runtime_; }
-  FORCE_INLINE int64_t contextId() const { return context_id_; }
+  FORCE_INLINE double contextId() const { return context_id_; }
 
   void InitializeQuickJSObject() override;
 
@@ -73,7 +73,7 @@ class ScriptWrappable : public GarbageCollected<ScriptWrappable> {
   JSValue jsObject_{JS_NULL};
   JSContext* ctx_{nullptr};
   ExecutingContext* context_{nullptr};
-  int64_t context_id_;
+  double context_id_;
   JSRuntime* runtime_{nullptr};
   WebFValueStatus* status_block_{nullptr};
   friend class GCVisitor;
