+++ bridge/webf_bridge.cc
@@ -141,7 +141,7 @@ void evaluateScripts(void* page_,
   auto page = reinterpret_cast<webf::WebFPage*>(page_);
   Dart_PersistentHandle persistent_handle = Dart_NewPersistentHandle_DL(dart_handle);
   page->executingContext()->dartIsolateContext()->dispatcher()->PostToJs(
-      page->isDedicated(), page->contextId(), webf::WebFPage::EvaluateScriptsInternal, page_, code, code_len,
+      page->isDedicated(), static_cast<int32_t>(page->contextId()), webf::WebFPage::EvaluateScriptsInternal, page_, code, code_len,
       parsed_bytecodes, bytecode_len, bundleFilename, start_line, profile_id, persistent_handle, result_callback);
 }
 
@@ -161,7 +161,7 @@ void dumpQuickjsByteCode(void* page_,
   auto page = reinterpret_cast<webf::WebFPage*>(page_);
   Dart_PersistentHandle persistent_handle = Dart_NewPersistentHandle_DL(dart_handle);
   page->dartIsolateContext()->dispatcher()->PostToJs(
-      page->isDedicated(), page->contextId(), webf::WebFPage::DumpQuickJsByteCodeInternal, page, profile_id, code,
+      page->isDedicated(), static_cast<int32_t>(page->contextId()), webf::WebFPage::DumpQuickJsByteCodeInternal, page, profile_id, code,
       code_len, parsed_bytecodes, bytecode_len, url, persistent_handle, result_callback);
 }
 
@@ -176,7 +176,7 @@ void evaluateQuickjsByteCode(void* page_,
 #endif
   auto page = reinterpret_cast<webf::WebFPage*>(page_);
   Dart_PersistentHandle persistent_handle = Dart_NewPersistentHandle_DL(dart_handle);
-  page->dartIsolateContext()->dispatcher()->PostToJs(page->isDedicated(), page->contextId(),
+  page->dartIsolateContext()->dispatcher()->PostToJs(page->isDedicated(), static_cast<int32_t>(page->contextId()),
                                                      webf::WebFPage::EvaluateQuickjsByteCodeInternal, page_, bytes,
                                                      byteLen, profile_id, persistent_handle, result_callback);
 }
@@ -193,7 +193,7 @@ void parseHTML(void* page_,
   auto page = reinterpret_cast<webf::WebFPage*>(page_);
   Dart_PersistentHandle persistent_handle = Dart_NewPersistentHandle_DL(dart_handle);
   page->executingContext()->dartIsolateContext()->dispatcher()->PostToJs(
-      page->isDedicated(), page->contextId(), webf::WebFPage::ParseHTMLInternal, page_, code, length, profile_id,
+      page->isDedicated(), static_cast<int32_t>(page->contextId()), webf::WebFPage::ParseHTMLInternal, page_, code, length, profile_id,
       persistent_handle, result_callback);
 }
 
@@ -240,7 +240,7 @@ void invokeModuleEvent(void* page_,
   auto dart_isolate_context = page->executingContext()->dartIsolateContext();
   auto is_dedicated = page->executingContext()->isDedicated();
   auto context_id = page->contextId();
-  dart_isolate_context->dispatcher()->PostToJs(is_dedicated, context_id, webf::WebFPage::InvokeModuleEventInternal,
+  dart_isolate_context->dispatcher()->PostToJs(is_dedicated, static_cast<int32_t>(context_id), webf::WebFPage::InvokeModuleEventInternal,
                                                page_, module, eventType, event, extra, persistent_handle,
                                                result_callback);
 }
@@ -251,7 +251,7 @@ void collectNativeProfileData(void* ptr, const char** data, uint32_t* len) {
 
   *data = static_cast<const char*>(webf::dart_malloc(sizeof(char) * result.size() + 1));
   memcpy((void*)*data, result.c_str(), sizeof(char) * result.size() + 1);
-  *len = result.size();
+  *len = static_cast<uint32_t>(result.size());
 }
 
 void clearNativeProfileData(void* ptr) {
